Main Google colab file
This the main file that contains the code for the Spam classifier. It uses Support Vector Machines to predict between a "Spam" and a "Ham". The "CountVectorizer" Library is used to convert a collection of text documents to a vector of term/token counts.  It also enables the ​pre-processing of text data prior to generating the vector representation. This functionality makes it a highly flexible feature representation module for text.